requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
dateparser==1.2.0
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
multidict==6.0.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyOpenSSL==24.1.0
python-binance==1.0.19
python-binance-aio==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
zope.interface==6.2
